+++ mlos_bench/mlos_bench/tests/services/remote/ssh/fixtures.py
@@ -9,7 +9,6 @@ Note: these are not in the conftest.py file because they are also used by remote
 """
 
 import os
-import sys
 import tempfile
 from collections.abc import Generator
 from subprocess import run
